move logic to provider
